Avances, Creaciond e ventas, Validacion de ventas, Funciones de buscar, Creaccion de objeto para mostrar los datos de las ventas

+++ 404-SDK-2022/Estructura/Struct-Ventas.cs
@@ -11,8 +11,8 @@ namespace DVStudio.SDK.Estructuras
         public string Fecha { get; set; }
         public Struct_Cliente Cliente { get; set; }
         public Struct_Usuarios Empleado { get; set; }
-        public int IVS { get; set; }
-        public int Total { get; set; }
+        public double IVS { get; set; }
+        public double Total { get; set; }
     }
     public struct Productos
     {
